test(epub): extract semantic visual family
